[cli] Refactored cli command
